refactor: standardize test organization - move comparison_tests to tests/ dir and add TESTING.md guide
