+++ crates/vibesql-types/src/sql_value/mod.rs
@@ -1,7 +1,6 @@
 //! SQL Value runtime representation
 
 mod comparison;
-mod comparison_tests;
 mod hash;
 mod display;
